fix: nil value in `ask` function
When asking cursoragent with a prompt (used in CursorAgentSelection and CursorAgentBuffer) the command to run is retrieved from a `get()` method in `cursoragent.config` that does not exist. This change gets the command from the config in the state object, and defaults to `cursor-agent` if no other command is found.